Move auth related items to /modules/auth
