--- src/guards/gql-auth.guard.ts
@@ -1,11 +1,0 @@
-import { Injectable, ExecutionContext } from '@nestjs/common'
-import { GqlExecutionContext } from '@nestjs/graphql'
-import { AuthGuard } from '@nestjs/passport'
-
-@Injectable()
-export class GqlAuthGuard extends AuthGuard('jwt') {
-  getRequest(context: ExecutionContext) {
-    const ctx = GqlExecutionContext.create(context)
-    return ctx.getContext().req
-  }
-}
