+++ src/modules/auth/jwt.strategy.ts
@@ -2,10 +2,10 @@ import { Injectable, UnauthorizedException } from '@nestjs/common'
 import { ConfigService } from '@nestjs/config'
 import { PassportStrategy } from '@nestjs/passport'
 import { Strategy, ExtractJwt } from 'passport-jwt'
-import { User } from 'src/models/user.model'
-import { AuthService } from 'src/services/auth.service'
 
-import { JwtDto } from './dto/jwt.dto'
+import { User } from '../user/user.model'
+import { JwtDto } from './dto'
+import { AuthService } from './services'
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
