+++ src/modules/auth/models/auth.model.ts
@@ -1,7 +1,7 @@
 import { Field, ObjectType } from '@nestjs/graphql'
 
+import { User } from '../../user/user.model'
 import { Token } from './token.model'
-import { User } from './user.model'
 
 @ObjectType()
 export class Auth extends Token {
