+++ src/modules/auth/models/index.ts
@@ -1,0 +1,2 @@
+export * from './auth.model'
+export * from './token.model'
