+++ src/modules/auth/models/token.model.ts
