+++ src/modules/auth/services/index.ts
@@ -1,0 +1,2 @@
+export * from './auth.service'
+export * from './password.service'
